+++ src/Forms/UI/views/fields/slider.blade.php
@@ -4,9 +4,9 @@
     name="{{ $getName($locale ?? null) }}"
     placeholder="{{ $getPlaceholder($locale ?? null) }}"
     value="{{ $getActiveValue($locale ?? null) }}"
-    min="{{ $getMin() ?? null }}"
-    max="{{ $getMax() ?? null }}"
-    step="{{ $getStep() ?? null }}"
+    :min="$getMin()"
+    :max="$getMax()"
+    :step="$getStep()"
     :autofocus="$hasAutofocus()"
     :attributes="$attributes
             ->merge($getCustomAttributes())