+++ src/Forms/UI/views/previews/fields/datetime.blade.php
@@ -1,6 +1,6 @@
 <p class="body body-dark wrap-anywhere">
     @if (! is_null($value = $getValueOrFallback($locale ?? null)) && $value !== '')
-        {{ rescue(fn () => \Carbon\Carbon::parse($value)->format('d/m/Y'), $value, false) }}
+        {{ rescue(fn () => \Carbon\Carbon::parse($value)->format('d/m/Y H:i'), $value, false) }}
     @else
         ...
     @endif